Created projects from scratch.

+++ SurveyMonkeyiOSBindings/Structs.cs
@@ -1,19 +1,14 @@
 ﻿using System;
-using UIKit;
-using Foundation;
-using ObjCRuntime;
-using CoreGraphics;
-
-
 
 namespace SurveyMonkeyiOSBindings
 {
-	/*
+
 	[Native]
-	public enum SMCompletionStatus : nuint
+	public enum SMCompletionStatus : ulong
 	{
 		PartiallyComplete,
 		Complete
 	}
-*/
+
 }
+
